functions from main moved to utils
