+++ backend/models.py
@@ -1,7 +1,5 @@
 from typing import Optional
 from sqlmodel import Field, SQLModel, Field, create_engine
-from pydantic import validator
-from utils import string_to_datetime
 
 con_str = f"sqlite:///database.sqlite"
 engine = create_engine(con_str, echo=True)
